fixed coordinates in database (8 decimals) and fixed controller file size anomaly

+++ database/migrations/2024_08_29_083312_create_apartments_table.php
@@ -21,8 +21,8 @@ return new class extends Migration
             $table->smallInteger('bathrooms');
             $table->smallInteger('dimension_mq');
             $table->string('image', 255);
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->decimal('latitude', 10, 8);
+            $table->decimal('longitude', 10, 8);
             $table->string('address_full');
             $table->boolean('is_visible');
             $table->timestamps();
